requirements.txt
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tkinter>=8.6
mlflow>=2.8.0 